+++ tests/crawlers/test_cases/fapello_com.py
@@ -1,0 +1,45 @@
+DOMAIN = "fapello.com"
+TEST_CASES = [
+    {
+        "url": "https://fapello.com/alessandra-ambrosio/6124",
+        "results": [
+            {
+                "url": "https://fapello.com/content/a/l/alessandra-ambrosio/7000/alessandra-ambrosio_6124.jpg",
+                "filename": "alessandra-ambrosio_6124.jpg",
+                "debrid_link": None,
+                "original_filename": "alessandra-ambrosio_6124.jpg",
+                "referer": "https://fapello.com/alessandra-ambrosio/6124",
+                "album_id": None,
+                "uploaded_at": None,
+                "download_folder": "re:alessandra-ambrosio (Fapello.com)",
+            }
+        ],
+        "count": 1,
+    },
+    {
+        "url": "https://fapello.com/soyloruga/54",
+        "results": [
+            {
+                "url": "https://cdn.fapello.com/content/s/o/soyloruga/1000/soyloruga_0054.mp4",
+                "filename": "soyloruga_0054.mp4",
+                "debrid_link": None,
+                "original_filename": "soyloruga_0054.mp4",
+                "referer": "https://fapello.com/soyloruga/54",
+                "album_id": None,
+                "uploaded_at": None,
+                "download_folder": "re:soyloruga (Fapello.com)",
+            }
+        ],
+        "count": 1,
+    },
+    {
+        "url": "https://fapello.com/soyloruga",
+        "results": [
+            {
+                "url": "ANY",
+                "download_folder": "re:soyloruga (Fapello.com)",
+            },
+        ],
+        "count": range(53, 60),
+    },
+]